+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 * Reverse function added.
 * Return function added.
 * Provision return fixed, Provision function changed accordingly.
+* BIN Info function added.